requirements.txt
fastapi
mcrcon
uvicorn
python-dotenv